Moved API logic into controllers
