+++ cmd/api/main.go
@@ -39,7 +39,11 @@ func main() {
 	playersRepository := players.NewPlayerRepository(dbHandler)
 	playersManager := players.NewManager(playersRepository)
 
-	err = api.SetUp(playersManager, gamesManager, seasonManager).Run(cfg.ApiHost)
+	gamesController := api.NewGamesController(gamesManager, seasonManager, playersManager)
+	seasonsController := api.NewSeasonsController(seasonManager)
+	playersController := api.NewPlayersController(playersManager, gamesManager)
+
+	err = api.SetUp(gamesController, seasonsController, playersController).Run(cfg.ApiHost)
 	if err != nil {
 		log.Fatal(err)
 	}
